[RFC]/MINOR: connection: Add server name to proxy protocol v2 header.
If the client provides the server name it intends to connect to, per RFC3546, Section 3.1. Server Name Indication, this patch will pass the server name onto the backend server as part of the proxy protocol v2 header.

The patch defines the new SSL subtype PP2_TYPE_SSL_SNI and the corresponding flag PP2_CLIENT_SNI to accomplish this in an additional TLV.